model: Register realm_user_groups during initial data fetch.
We register realm_user_groups event type during initial data fetch
to get information regarding user_groups in realm.

We pass the information to a private method to extract
and store the information in index to be used later.

Tests amended for conftest and TestModel.

Tests added for _group_info_from_*