helper: Avoid duplicate quoting using shlex for notifications.
Shlex used to quote the text using single quotes which caused
the shlex.split in subprocess.run to split at single quotes
and make the command have multiple parameters such as -->

['osascript', '-e', 'display notification "text',
'text" with title "Zulip Community"']

This caused the command to fail and the notification did not work.
This patch avoides the extra quoting using shlex.